backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.0
alembic>=1.16.0
psycopg2-binary>=2.9.0
python-dotenv>=1.1.0
pydantic[email]>=2.11.0
pydantic-settings>=2.10.0
passlib[bcrypt]>=1.7.0
pyjwt>=2.10.0
httpx>=0.28.0
boto3>=1.38.0
python-multipart>=0.0.20
apscheduler>=3.11.0
stripe>=12.2.0
razorpay>=1.4.0
starlette>=0.46.0
